# services/gateway-svc/api/rest_oapi/openapi.yaml
openapi: 3.0.3
info:
  title: gateway-svc API
  version: 0.1.0
  description: |
    Edge REST API for UmrohOS. gateway-svc fronts every backend service and is
    the only outward-facing HTTP surface clients (web app, mobile app, CLI)
    should call.

    Scaffold scope: 12 endpoints —

      - 2 own probes: /system/live, /system/ready
      - 10 per-backend liveness proxies: /v1/<svc>/system/live, one per
        backend in the umrohos service map. Used by the web app's status page.

    Real per-route proxies (auth, packages, bookings, ...) land alongside
    each backend's first feature work.
servers:
  - url: /
    description: Base path
paths:
...
  /v1/iam/system/diagnostics/db-tx:
    get:
      tags: [Proxy / iam]
      summary: Proxy iam-svc DB transaction diagnostic
      description: |
        Traced cross-service verification path. Proxies to iam-svc's
        state-mutating `/system/diagnostics/db-tx` diagnostic, which inserts a
        row into `public.diagnostics` inside a WithTx and echoes the message
        back. Used by the `S0-J-05` observability acceptance check: a single
        request through this endpoint must produce one trace in Tempo whose
        spans span gateway-svc and iam-svc, and matching `trace_id`-tagged log
        lines in Loki for both containers.

        Pilot-scope endpoint; not for health/readiness probes. Gate with auth
        / a header secret before promoting past pilot.
      operationId: getIamSystemDbTxDiagnostic
      parameters:
        - name: message
          in: query
          required: false
          schema:
            type: string
            default: "no message"
          description: Message to store and echo back.
      responses:
        "200":
          description: Upstream OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DbTxDiagnosticResponse" }
        "502":
          description: Upstream call failed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }

  # /v1/catalog/system/live removed in S1-E-11: catalog-svc is now gRPC-only.
  # /v1/iam/system/live removed in S1-E-12: iam-svc is now gRPC-only.
  # IAM client-facing auth routes below replace the former iam-svc REST surface.

  # ── IAM Auth routes (BL-IAM-018 / S1-E-12) ─────────────────────────────
  # gateway-svc proxies these to iam-svc via iam_grpc_adapter per ADR 0009.
  # Request/response shapes mirror iam-svc's former REST surface exactly so
  # existing clients (web app, mobile) do not need to change.
...
  /v1/auth/refresh:
    post:
      tags: [IAM / Auth]
      summary: Rotate refresh token — issue a new access + refresh pair
      operationId: refreshSession
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RefreshSessionRequest" }
      responses:
        "200":
          description: New token pair issued
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RefreshSessionResponse" }
        "400":
          description: Malformed request body
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "401":
          description: Refresh token unknown, expired, or revoked (replay)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  /v1/me/2fa/enroll:
    post:
      tags: [IAM / Me]
      summary: Begin TOTP enrollment
      operationId: enrollTOTP
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Enrollment started
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EnrollTOTPResponse" }
        "401":
          description: Missing or invalid bearer token
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "409":
          description: TOTP already verified for this user
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /v1/me/2fa/verify:
    post:
      tags: [IAM / Me]
      summary: Verify TOTP code and mark enrollment complete
      operationId: verifyTOTP
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/VerifyTOTPRequest" }
      responses:
        "200":
          description: Verification successful
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VerifyTOTPResponse" }
        "400":
          description: Malformed request body
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "401":
          description: Invalid TOTP code or missing/invalid bearer token
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /v1/users/{id}/suspend:
    post:
      tags: [IAM / Admin]
      summary: Suspend a user and revoke every active session
      operationId: suspendUser
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
          description: Target user UUID.
      responses:
        "200":
          description: User suspended
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SuspendUserResponse" }
        "400":
          description: Malformed user id OR self-suspend
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "401":
          description: Missing or invalid bearer token
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "403":
          description: Bearer's user lacks iam.users/suspend/global permission
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "404":
          description: Target user not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }

  # ── Public catalog read (BL-GTW-002 / S1-E-10) ─────────────────────────
  # Gateway REST → catalog-svc gRPC via catalog_grpc_adapter. Responses
  # mirror services/catalog-svc/api/rest_oapi/openapi.yaml § Catalog
  # shape-for-shape so e2e specs passing against either port yield
  # identical bodies. After G7 (BL-REFACTOR-001) the catalog REST port
  # goes away and these gateway routes become the only client-facing
  # surface.
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: PASETO
      description: |
        Single-point edge authentication per ADR 0009 / F1-W7. Clients send
        `Authorization: Bearer <access_token>` on any protected route.
        gateway-svc validates the token once against `iam-svc.ValidateToken`
        (gRPC); backends do NOT re-validate. Failure modes: missing/malformed
        → 401 UNAUTHORIZED; iam-svc rejects → 401 UNAUTHORIZED; iam-svc
        unreachable → 502 SERVICE_UNAVAILABLE (fail-closed).

        No route is currently tagged with this scheme — first consumer lands
        with BL-IAM-018 / S1-E-12 when iam auth REST routes move to gateway.
  schemas:
...
    DbTxDiagnosticResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [diagnostic_id, message]
          properties:
            diagnostic_id:
              type: integer
              format: int64
              description: ID of the row inserted by iam-svc.
            message:
              type: string
              description: Echo of the message received from the client.
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:    { type: string, example: BAD_GATEWAY }
            message: { type: string, example: upstream service unavailable }

    # ── Catalog error envelope (BL-GTW-002 / S1-E-10) ──────────────────
    # Mirror of services/catalog-svc/api/rest_oapi/openapi.yaml
    # ErrorResponse: snake_case code, id-ID message, optional trace_id.
    # Catalog handlers on gateway write this directly (bypassing the
    # default gateway error middleware that emits UPPER_SNAKE codes).
...
    UserProfile:
      type: object
      required: [user_id, email, name, branch_id, status]
      properties:
        user_id:    { type: string, format: uuid }
        email:      { type: string, format: email }
        name:       { type: string }
        branch_id:  { type: string, format: uuid }
        status:
          type: string
          enum: [active, suspended, pending]
...
    RefreshSessionRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token: { type: string }
    RefreshSessionResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [access_token, refresh_token, access_expires_at, refresh_expires_at]
          properties:
            access_token:       { type: string }
            refresh_token:      { type: string }
            access_expires_at:  { type: string, format: date-time }
            refresh_expires_at: { type: string, format: date-time }
...
    EnrollTOTPResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [secret, otpauth_url]
          properties:
            secret:      { type: string, description: "Base32 TOTP secret — display to user once." }
            otpauth_url: { type: string, description: "otpauth:// URI for QR-code generation." }
    VerifyTOTPRequest:
      type: object
      required: [code]
      properties:
        code: { type: string, description: "6-digit TOTP code from the authenticator app." }
    VerifyTOTPResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [verified_at]
          properties:
            verified_at: { type: string, format: date-time }
    SuspendUserResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [user]
          properties:
            user: { $ref: "#/components/schemas/UserProfile" }

    # ---------------------------------------------------------------------------
    # Staff catalog write schemas (BL-CAT-014 / S1-E-07)
    # ---------------------------------------------------------------------------
...